fixed review comments.
