+++ Bird-Modules/Sources/NewFlashcardFeature/NewFlashcardView.swift
@@ -23,7 +23,7 @@ public struct NewFlashcardView: View {
     @StateObject private var frontContext = RichTextContext()
     @StateObject private var backContext = RichTextContext()
     
-    @FocusState private var focus: NewFlashcardFocus?
+//    @FocusState private var focus: NewFlashcardFocus?
     
     @Environment(\.dismiss) private var dismiss
     
@@ -88,9 +88,9 @@ public struct NewFlashcardView: View {
                     customAlert()
                 }
                 .toolbar {
-                    ToolbarItemGroup(placement: .keyboard) {
-                        customrightToolbarItemGroup
-                    }
+//                    ToolbarItemGroup(placement: .keyboard) {
+//                        customrightToolbarItemGroup
+//                    }
                     ToolbarItem(placement: .navigationBarTrailing) {
                         customNavigationToolbar
                     }
@@ -119,7 +119,6 @@ public struct NewFlashcardView: View {
                 }     
             }
         }
-        .interactiveDismissDisabled(focus != nil ? true : false)
     }
     
     @ViewBuilder
@@ -172,35 +171,35 @@ public struct NewFlashcardView: View {
         }
     }
     
-    @ViewBuilder
-    private var customrightToolbarItemGroup: some View {
-        Spacer()
-        Button {
-            if focus == .back {
-                focus = .front
-            }
-        } label: {
-            Image(systemName: "chevron.up")
-        }
-        .disabled(focus == .front)
-        .accessibilityLabel(focus == .front ? NSLocalizedString("moveup_focus_disabled", bundle: .module, comment: "") : NSLocalizedString("moveup_focus", bundle: .module, comment: ""))
-        
-        
-        Button {
-            if focus == .front {
-                focus = .back
-            }
-        } label: {
-            Image(systemName: "chevron.down")
-        }
-        .disabled(focus == .back)
-        .accessibilityLabel(focus == .back ? NSLocalizedString("down_focus_disabled", bundle: .module, comment: "") : NSLocalizedString("down_focus", bundle: .module, comment: ""))
-        
-        Button(NSLocalizedString("feito", bundle: .module, comment: "")) {
-            focus = nil
-        }
-        .accessibilityLabel(Text("botao_feito", bundle: .module))
-    }
+//    @ViewBuilder
+//    private var customrightToolbarItemGroup: some View {
+//        Spacer()
+//        Button {
+//            if focus == .back {
+//                focus = .front
+//            }
+//        } label: {
+//            Image(systemName: "chevron.up")
+//        }
+//        .disabled(focus == .front)
+//        .accessibilityLabel(focus == .front ? NSLocalizedString("moveup_focus_disabled", bundle: .module, comment: "") : NSLocalizedString("moveup_focus", bundle: .module, comment: ""))
+//
+//
+//        Button {
+//            if focus == .front {
+//                focus = .back
+//            }
+//        } label: {
+//            Image(systemName: "chevron.down")
+//        }
+//        .disabled(focus == .back)
+//        .accessibilityLabel(focus == .back ? NSLocalizedString("down_focus_disabled", bundle: .module, comment: "") : NSLocalizedString("down_focus", bundle: .module, comment: ""))
+//
+//        Button(NSLocalizedString("feito", bundle: .module, comment: "")) {
+//            focus = nil
+//        }
+//        .accessibilityLabel(Text("botao_feito", bundle: .module))
+//    }
     
     @ViewBuilder
     private var customNavigationToolbar: some View {
